Move tests for /utils/script/evaluate to separate class
